requirements.txt
boto3>=1.29.0
botocore>=1.32.0
moviepy>=1.0.3
python-dotenv>=1.0.0
requests>=2.31.0

# Optional packages for enhanced functionality
nltk>=3.8.1
spacy>=3.7.0
transformers>=4.35.0
torch>=2.0.0

# Development and testing
pytest>=7.4.0
pytest-asyncio>=0.21.0
black>=23.0.0
flake8>=6.0.0